코드리뷰 반영
- https://github.com/next-step/swift-racingcar/pull/17/files/9494f418d30b66d768ac62a8ee2759be0aad6395#r741569619
