+++ RacingCarTest/ExpressionParserTests.swift
@@ -43,23 +43,3 @@ class ExpressionParserTests: XCTestCase {
         XCTAssertThrowsError(try expressionParser.parse(expression: expression))
     }
 }
-
-protocol Parserable {
-    
-    func parse(expression: String) throws -> [String]
-}
-
-class ExpressionParser: Parserable {
-    
-    enum ParsingError: Error {
-        
-        case emptyInput
-    }
-    
-    func parse(expression: String) throws -> [String] {
-        guard !expression.isEmpty
-        else { throw ParsingError.emptyInput }
-        
-        return expression.components(separatedBy: " ")
-    }
-}
